refactor evm fallback, add method for grabbing top tokens on EVM chains
